Creates profile
 By signing up user in cognito and cassandra calling sign up API
 Save security questions by calling the security Questions API(First gets JWt token
  in eash step it will return approrate http status code
- Added Validation for password, username.phone and email
- Added validationfor contract account and Bp matching
- Added validation to check username and contarct account

Related work items: #114, #2113, #2114

+++ src/V1/Models/SecurityQuestionResponse.cs
@@ -6,7 +6,7 @@ namespace PSE.Customer.V1.Models
     public class SecurityQuestionResponse: IAPIResponse
     {
         [Required]
-        public int Sequence { get; set; }
+        public short Sequence { get; set; }
         [Required]
         public string Question { get; set; }
         [Required]
